Move shared stats and sender helpers
