+++ apps/web/app/api/user/categorize/senders/uncategorized/route.ts
@@ -1,7 +1,7 @@
 import { NextResponse } from "next/server";
 import { withEmailAccount } from "@/utils/middleware";
-import { getUncategorizedSenders } from "@/app/api/user/categorize/senders/uncategorized/get-uncategorized-senders";
-import type { Sender } from "@/app/api/user/categorize/senders/batch/handle-batch-validation";
+import { getUncategorizedSenders } from "@/utils/categorize/senders/get-uncategorized-senders";
+import type { Sender } from "@/utils/categorize/senders/batch-validation";
 
 export type UncategorizedSendersResponse = {
   uncategorizedSenders: Sender[];
