+++ apps/web/app/api/user/stats/response-time/route.ts
@@ -1,7 +1,7 @@
 import { NextResponse } from "next/server";
 import { withEmailProvider } from "@/utils/middleware";
-import { getResponseTimeStats } from "./controller";
-import { responseTimeQuerySchema } from "@/app/api/user/stats/response-time/validation";
+import { getResponseTimeStats } from "@/utils/stats/response-time/controller";
+import { responseTimeQuerySchema } from "@/utils/stats/response-time/validation";
 
 export const GET = withEmailProvider("response-time-stats", async (request) => {
   const { searchParams } = new URL(request.url);
