+++ apps/web/utils/categorize/senders/batch-validation.ts
