+++ apps/web/utils/categorize/senders/start-bulk-categorization.ts
@@ -13,7 +13,7 @@ import {
   deleteEmptyCategorizeSendersQueues,
   publishToAiCategorizeSendersQueue,
 } from "@/utils/upstash/categorize-senders";
-import { getUncategorizedSenders } from "@/app/api/user/categorize/senders/uncategorized/get-uncategorized-senders";
+import { getUncategorizedSenders } from "@/utils/categorize/senders/get-uncategorized-senders";
 import { loadEmails } from "@/utils/actions/stats-loading";
 
 const CATEGORIZE_SYNC_CHUNK_PAGES = 5;
